Develop

+++ banking-service/src/main/java/com/finpilot/banking/service/AccountService.java
@@ -2,11 +2,13 @@ package com.finpilot.banking.service;
 
 import com.finpilot.banking.entity.Account;
 
+import java.math.BigDecimal;
+
 public interface AccountService {
 
     Account createAccount(Account account);
 
     Account getAccount(Long id);
 
-    Double getBalance(Long id);
+    BigDecimal getBalance(Long id);
 }
